Unit Price, Unit 이름 다국어 적용안되는 오류 수정 독일어 교정 결과 반영 다국어 누락분 Wallet Field Style 추가

+++ AppBox3/Common/Passcode/A3PasscodeViewController.m
@@ -741,7 +741,7 @@ static NSInteger const kMaxNumberOfAllowedFailedAttempts = 10;
 
 	if (_failedAttempts == 1) _failedAttemptLabel.text = NSLocalizedString(@"1 Passcode Failed Attempt", @"");
 	else {
-		_failedAttemptLabel.text = [NSString stringWithFormat: NSLocalizedString(@"%i Passcode Failed Attempts", @""), _failedAttempts];
+		_failedAttemptLabel.text = [NSString stringWithFormat: NSLocalizedString(@"%li Passcode Failed Attempts", @""), (long)_failedAttempts];
 	}
 	_failedAttemptLabel.layer.cornerRadius = kFailedAttemptLabelHeight * 0.65f;
 	_failedAttemptLabel.hidden = NO;
